requirements.txt
openai
google-cloud-speech
google-cloud-texttospeech
pyaudio
pygame
datetime
webrtcvad
